separate template rendering out from __main__
